Agentic Workflows/Youtube Analysis — Generic/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
youtube-transcript-api>=0.6.2
anthropic>=0.34.0
python-dotenv>=1.0.0
